change the pc value gdb reads from qemu
Currently we pass `skiboot.lid` to QEMU, but try to debug the same with
`skiboot.elf` to GDB.

There is a mismatch between the addresses for same instructions between
these files. In case of QEMU, most of the skiboot code is
offset/relocated by `0x30000000`.

So the code and instruction we see at `0x22af` in GDB (in skiboot.elf),
is actually at `0x300022af` in QEMU (running skiboot.lid).

Even with this mismatch, we are able to get assembly instructions from
QEMU, since that we are reading from QEMU's memory, but anything more
complex doesn't work.

1. Instruction to source code mapping doesn't work.

Fix this by modifying the registers that GDB is getting from QEMU.
So, if QEMU returns that it's PC is `0x300022af`, we will make GDB think
that it is `0x22af`, thus things depending on address to source code
mapping starts working in GDB.

2. Setting breakpoints at functions doesn't work

This is since GDB will try to get the instruction address of the
function using debug info from skiboot.elf, which is different from
runtime address as explained in beginning.

Fix it by modifying the breakpoint address that gdb will be sending to
QEMU.
Do note, that this causes a problem, that the code doesn't progress at
all due to SIGTRAP, just do a `stepi`, to force some progress, then we
can use whatever we want to use such as `continue` command

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>